demo resources: move demo code to the demo folder
